test(core): make Distribution slot test portable across toolchains
